backend/app/requirements.txt
fastapi
uvicorn
pydantic
dashscope
python-dotenv 